+++ saleor/graphql/meta/tests/test_meta_mutations.py
@@ -417,7 +417,7 @@ def test_add_public_metadata_for_order_by_token(api_client, order):
 
     # when
     response = execute_update_public_metadata_for_item(
-        api_client, None, order.token, "Order"
+        api_client, None, order.id, "Order"
     )
 
     # then
@@ -447,7 +447,7 @@ def test_add_public_metadata_for_draft_order_by_token(api_client, draft_order):
 
     # when
     response = execute_update_public_metadata_for_item(
-        api_client, None, draft_order.token, "Order"
+        api_client, None, draft_order.id, "Order"
     )
 
     # then
@@ -1225,7 +1225,7 @@ def test_delete_public_metadata_for_order_by_token(api_client, order):
 
     # when
     response = execute_clear_public_metadata_for_item(
-        api_client, None, order.token, "Order"
+        api_client, None, order.id, "Order"
     )
 
     # then
@@ -1259,7 +1259,7 @@ def test_delete_public_metadata_for_draft_order_by_token(api_client, draft_order
 
     # when
     response = execute_clear_public_metadata_for_item(
-        api_client, None, draft_order.token, "Order"
+        api_client, None, draft_order.id, "Order"
     )
 
     # then
@@ -2032,7 +2032,7 @@ def test_add_private_metadata_for_order_by_token(
 
     # when
     response = execute_update_private_metadata_for_item(
-        staff_api_client, permission_manage_orders, order.token, "Order"
+        staff_api_client, permission_manage_orders, order.id, "Order"
     )
 
     # then
@@ -2066,7 +2066,7 @@ def test_add_private_metadata_for_draft_order_by_token(
 
     # when
     response = execute_update_private_metadata_for_item(
-        staff_api_client, permission_manage_orders, draft_order.token, "Order"
+        staff_api_client, permission_manage_orders, draft_order.id, "Order"
     )
 
     # then
@@ -2885,7 +2885,7 @@ def test_delete_private_metadata_for_order_by_token(
 
     # when
     response = execute_clear_private_metadata_for_item(
-        staff_api_client, permission_manage_orders, order.token, "Order"
+        staff_api_client, permission_manage_orders, order.id, "Order"
     )
 
     # then
@@ -2923,7 +2923,7 @@ def test_delete_private_metadata_for_draft_order_by_token(
 
     # when
     response = execute_clear_private_metadata_for_item(
-        staff_api_client, permission_manage_orders, draft_order.token, "Order"
+        staff_api_client, permission_manage_orders, draft_order.id, "Order"
     )
 
     # then