+++ saleor/graphql/webhook/mutations.py
@@ -6,7 +6,7 @@ from ...webhook import models
 from ...webhook.error_codes import WebhookErrorCode
 from ..core.descriptions import DEPRECATED_IN_3X_INPUT
 from ..core.mutations import BaseMutation, ModelDeleteMutation, ModelMutation
-from ..core.types.common import WebhookError
+from ..core.types import NonNullList, WebhookError
 from . import enums
 from .types import EventDelivery, Webhook
 
@@ -14,19 +14,19 @@ from .types import EventDelivery, Webhook
 class WebhookCreateInput(graphene.InputObjectType):
     name = graphene.String(description="The name of the webhook.", required=False)
     target_url = graphene.String(description="The url to receive the payload.")
-    events = graphene.List(
+    events = NonNullList(
         enums.WebhookEventTypeEnum,
         description=(
             f"The events that webhook wants to subscribe. {DEPRECATED_IN_3X_INPUT} "
             "Use `asyncEvents` or `syncEvents` instead."
         ),
     )
-    async_events = graphene.List(
-        graphene.NonNull(enums.WebhookEventTypeAsyncEnum),
+    async_events = NonNullList(
+        enums.WebhookEventTypeAsyncEnum,
         description="The asynchronous events that webhook wants to subscribe.",
     )
-    sync_events = graphene.List(
-        graphene.NonNull(enums.WebhookEventTypeSyncEnum),
+    sync_events = NonNullList(
+        enums.WebhookEventTypeSyncEnum,
         description="The synchronous events that webhook wants to subscribe.",
     )
     app = graphene.ID(
@@ -122,7 +122,7 @@ class WebhookUpdateInput(graphene.InputObjectType):
     target_url = graphene.String(
         description="The url to receive the payload.", required=False
     )
-    events = graphene.List(
+    events = NonNullList(
         enums.WebhookEventTypeEnum,
         description=(
             f"The events that webhook wants to subscribe. {DEPRECATED_IN_3X_INPUT} "
@@ -130,13 +130,13 @@ class WebhookUpdateInput(graphene.InputObjectType):
         ),
         required=False,
     )
-    async_events = graphene.List(
-        graphene.NonNull(enums.WebhookEventTypeAsyncEnum),
+    async_events = NonNullList(
+        enums.WebhookEventTypeAsyncEnum,
         description="The asynchronous events that webhook wants to subscribe.",
         required=False,
     )
-    sync_events = graphene.List(
-        graphene.NonNull(enums.WebhookEventTypeSyncEnum),
+    sync_events = NonNullList(
+        enums.WebhookEventTypeSyncEnum,
         description="The synchronous events that webhook wants to subscribe.",
         required=False,
     )