+++ saleor/graphql/webhook/types.py
@@ -11,7 +11,7 @@ from ..core.connection import (
 )
 from ..core.descriptions import DEPRECATED_IN_3X_FIELD
 from ..core.fields import FilterConnectionField
-from ..core.types import ModelObjectType
+from ..core.types import ModelObjectType, NonNullList
 from ..webhook.enums import EventDeliveryStatusEnum, WebhookEventTypeEnum
 from ..webhook.filters import EventDeliveryFilterInput
 from ..webhook.sorters import (
@@ -138,21 +138,21 @@ class EventDeliveryCountableConnection(CountableConnection):
 class Webhook(ModelObjectType):
     id = graphene.GlobalID(required=True)
     name = graphene.String(required=True)
-    events = graphene.List(
-        graphene.NonNull(WebhookEvent),
+    events = NonNullList(
+        WebhookEvent,
         description="List of webhook events.",
         deprecation_reason=(
             f"{DEPRECATED_IN_3X_FIELD} Use `asyncEvents` or `syncEvents` instead."
         ),
         required=True,
     )
-    sync_events = graphene.List(
-        graphene.NonNull(WebhookEventSync),
+    sync_events = NonNullList(
+        WebhookEventSync,
         description="List of synchronous webhook events.",
         required=True,
     )
-    async_events = graphene.List(
-        graphene.NonNull(WebhookEventAsync),
+    async_events = NonNullList(
+        WebhookEventAsync,
         description="List of asynchronous webhook events.",
         required=True,
     )