+++ saleor/plugins/webhook/tests/test_shippping_exclude_cache.py
@@ -110,7 +110,7 @@ def test_excluded_shipping_methods_for_order_stores_in_cache_when_empty(
     # then
     assert mocked_webhook.called
 
-    expected_cache_key = CACHE_EXCLUDED_SHIPPING_KEY + order_with_lines.token
+    expected_cache_key = CACHE_EXCLUDED_SHIPPING_KEY + str(order_with_lines.id)
 
     expected_excluded_shipping_method = [{"id": "1", "reason": webhook_reason}]
 
@@ -174,7 +174,7 @@ def test_excluded_shipping_methods_for_order_stores_in_cache_when_payload_is_dif
     # then
     assert mocked_webhook.called
 
-    expected_cache_key = CACHE_EXCLUDED_SHIPPING_KEY + order_with_lines.token
+    expected_cache_key = CACHE_EXCLUDED_SHIPPING_KEY + str(order_with_lines.id)
 
     expected_excluded_shipping_method = [{"id": "1", "reason": webhook_reason}]